+++ lists/mirror-tv/Contact.js
@@ -62,8 +62,8 @@ module.exports = {
             label: 'Twitter',
             type: Url,
         },
-        instatgram: {
-            label: 'Instatgram',
+        instagram: {
+            label: 'Instagram',
             type: Url,
         },
         bio: {